Fix edge case where robot wont grab
The Robot wouldn't grab if the token was above the robot while the
marker was below, this is because the bearings of the robot and
token would read -pi and pi. The offset would then read as 2pi,
instead of 0 and therefore the token wouldn't be grabbed.

This is fixed by adding an offset to wrap the total offset to
never be over pi.